pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "splitsquash"
version = "0.0.3"
requires-python = ">=3.12"
authors = [
  {name = "Charlie Bareham", email = "barehamcharlie@gmail.com"},
]
description = "An editor for performing interactive rebases with extra features."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
classifiers = [
  "Programming Language :: Python"
]
dependencies = [
    "GitPython",
    "textual"
]

[project.urls]
Repository = "https://github.com/me/spam.git"

[project.scripts]
splitsquash = "splitsquash.scripts.editor:main"
ss-edit-rebase-item = "splitsquash.scripts.edit_rebase_item:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
splitsquash = ["styles/*.tcss"]
